+++ Classes/WAOperationsManager.m
@@ -26,7 +26,7 @@ static WAOperationsManager *sharedMyManager = nil;
 - (id)retain {
 	return self;
 }
-- (unsigned)retainCount {
+- (NSUInteger)retainCount {
 	return UINT_MAX; //denotes an object that cannot be released
 }
 - (oneway void)release {
